+++ packages/implementations/react/src/atomic/Input/src/components/Search/Search.tsx
@@ -20,6 +20,7 @@ const InputSearch = forwardRef<HTMLInputElement, _InputSearchProps>(
       size = "medium",
       border = true,
       divider = true,
+      disabled,
       appearance,
       ...rest
     }: _InputSearchProps,
@@ -38,6 +39,7 @@ const InputSearch = forwardRef<HTMLInputElement, _InputSearchProps>(
         <input
           {...rest}
           ref={inputRef}
+          disabled={disabled}
           type="search"
           className={input.classnames.input({ size, divider })}
         />
@@ -48,6 +50,7 @@ const InputSearch = forwardRef<HTMLInputElement, _InputSearchProps>(
           appearance={appearance}
           size={size}
           divider={divider}
+          disabled={disabled}
         >
           <Icon
             color="$currentColor"