+++ packages/implementations/react/src/atomic/Input/CHANGELOG.md
@@ -2,11 +2,20 @@
 
 The Input component allows users to enter text into a UI. They typically appear on forms and dialog boxes.
 
-## 2024-06-21 `1.1.2`
+## 2024-06-25 `1.1.3`
 
 #### 🐛 Bug fixes
 
 - Added the size prop for the `container` class of the `Input.Password` and `Input.Search`. ([#60](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/60) by [@vinicius.osmainchi](https://git.rarolabs.com.br/vinicius.osmainchi))
+
+## 2024-06-25 `1.1.2`
+
+#### 🐛 Bug fixes
+
+- Adjusted colors of the disabled state of the `Input` component. ([#62](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/62) by [@junior](https://git.rarolabs.com.br/junior))
+- Adjusted the height of the component `Input`. ([#62](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/62) by [@junior](https://git.rarolabs.com.br/junior))
+- Adjusted the height of the component `Input.Password`. ([#62](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/62) by [@junior](https://git.rarolabs.com.br/junior))
+- Adjusted the height of the component `Input.search`. ([#62](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/62) by [@junior](https://git.rarolabs.com.br/junior))
 
 ## 2024-06-17 `1.1.1`